// src/app/models/admin-models/model.about.ts
export class About {
  id: number;

  is_publish: number;

  ua_title?: string;
  eng_title?: string;
  rus_title?: string;

  ua_main_title_number?: string;
  eng_main_title_number?: string;
  rus_main_title_number?: string;

  ua_main_text_number?: string;
  eng_main_text_number?: string;
  rus_main_text_number?: string;

  ua_main_title_fact?: string;
  eng_main_title_fact?: string;
  rus_main_title_fact?: string;

  ua_main_text_fact?: string;
  eng_main_text_fact?: string;
  rus_main_text_fact?: string;

  ua_main_more?: string;
  eng_main_more?: string;
  rus_main_more?: string;

  number_field1?: string;
  number_field2?: string;
  number_field3?: string;

  number_ua_title1?: string;
  number_eng_title1?: string;
  number_rus_title1?: string;

  number_ua_text1?: string;
  number_eng_text1?: string;
  number_rus_text1?: string;

  number_ua_title2?: string;
  number_eng_title2?: string;
  number_rus_title2?: string;

  number_ua_text2?: string;
  number_eng_text2?: string;
  number_rus_text2?: string;

  number_ua_title3?: string;
  number_eng_title3?: string;
  number_rus_title3?: string;

  number_ua_text3?: string;
  number_eng_text3?: string;
  number_rus_text3?: string;

  number_photo1?: string;
  number_photo2?: string;

  ua_facts_field1?: string;
  eng_facts_field1?: string;
  rus_facts_field1?: string;

  ua_facts_field2?: string;
  eng_facts_field2?: string;
  rus_facts_field2?: string;

  ua_facts_field3?: string;
  eng_facts_field3?: string;
  rus_facts_field3?: string;

  ua_facts_field4?: string;
  eng_facts_field4?: string;
  rus_facts_field4?: string;

  facts_photo1?: string;
  facts_photo2?: string;

  createdAt?: string;
  updatedAt?: string;
}